requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.0.1
openai==2.6.1
psycopg2==2.9.11
sqlalchemy==2.0.30
pydantic==2.12.3
python-multipart==0.0.21
psycopg2-binary==2.9.11
pdfplumber==0.11.9
python-docx==1.2.0
pandas==2.2.2
openpyxl==3.1.2
python-pptx==1.0.2
chromadb==1.4.1
pdfplumber==0.11.9
pytesseract==0.3.13